ISL

Documentation for ISL.jl

ISL.AutoISLParams — Type

AutoISLParams

Hyperparameters for the method invariant_statistical_loss

@with_kw struct AutoISLParams
    samples::Int64 = 1000
    epochs::Int64 = 100
    η::Float64 = 1e-3
    max_k::Int64 = 10
    transform = Normal(0.0f0, 1.0f0)
end;

ISL.HyperParamsTS — Type

HyperParamsTS

Hyperparameters for the method ts_adaptative_block_learning

ISL.ISLParams — Type

ISLParams

Hyperparameters for the method adaptative_block_learning

@with_kw struct ISLParams
    samples::Int64 = 1000               # number of samples per histogram
    K::Int64 = 2                        # number of simulted observations
    epochs::Int64 = 100                 # number of epochs
    η::Float64 = 1e-3                   # learning rate
    transform = Normal(0.0f0, 1.0f0)    # transform to apply to the data
end;

ISL._sigmoid — Method

_sigmoid(ŷ::Matrix{T}, y::T) where {T<:AbstractFloat}

Calculate the sigmoid function centered at y.

Arguments

ŷ::Matrix{T}: The matrix of values to apply the sigmoid function to.

y::T: The center value around which the sigmoid function is centered.

Returns

A matrix of the same size as ŷ containing the sigmoid-transformed values.

This function calculates the sigmoid function for each element in the matrix ŷ centered at the value y. It applies a fast sigmoid transformation with a scaling factor of 10.0.

ISL.auto_invariant_statistical_loss — Method

`auto_invariant_statistical_loss(model, data, hparams)``

Custom loss function for the model.

This method gradually adapts K (starting from 2) up to max_k (inclusive). The value of K is chosen based on a simple two-sample test between the histogram associated with the obtained result and the uniform distribution.

To see the value of K used in the test, set the logger level to debug before executing.

#Arguments

model::Flux.Chain: is a Flux neuronal network model

data::Flux.DataLoader: is a loader Flux object

hparams::AutoAdaptativeHyperParams: is a AutoAdaptativeHyperParams object

ISL.convergence_to_uniform — Method

`convergence_to_uniform(aₖ)``

Test the convergence of the distributino of the window of the rv's Aₖ to a uniform distribution. It is implemented using a Chi-Square test.

ISL.generate_aₖ — Method

`generate_aₖ(ŷ, y)``

Calculate the values of the real observation y in each of the components of the approximate histogram with K bins.

Arguments

ŷ::Matrix{T}: A matrix of simulated observations (each column represents a different simulation).

y::T: The real data for which the one-step histogram is generated.

Returns

aₖ::Vector{Float}: A vector with the values of the real observation y in each of the components of the approximate histogram with K bins.

Details

The generate_aₖ function calculates the one-step histogram aₖ as the sum of the contribution of the observation to the subrogate histogram bins. It uses the function γ to calculate the contribution of each observation at each histogram bin. The final histogram is the sum of these contributions.

The formula for generating aₖ is as follows:

\[aₖ = ∑_{k=0}^K γ(ŷ, y, k) = ∑_{k=0}^K ∑_{i=1}^N ψₖ(ŷ, yᵢ)\]

ISL.get_window_of_Aₖ — Method

`get_window_of_Aₖ(model, target , K, n_samples)``

Generate a window of the rv's Aₖ for a given model and target function.

ISL.invariant_statistical_loss — Method

`invariant_statistical_loss(model, data, hparams)``

Custom loss function for the model. model is a Flux neuronal network model, data is a loader Flux object and hparams is a HyperParams object.

Arguments

nn_model::Flux.Chain: is a Flux neuronal network model

data::Flux.DataLoader: is a loader Flux object

hparams::HyperParams: is a HyperParams object

ISL.jensen_shannon_∇ — Method

`jensen_shannon_∇(aₖ)``

Jensen shannon difference between aₖ vector and uniform distribution vector.

ISL.scalar_diff — Method

`scalar_diff(q)`

Scalar difference between the vector representing our subrogate histogram and the uniform distribution vector.

\[loss = ||q-1/k+1||_{2} = ∑_{k=0}^K (qₖ - 1/K+1)^2\]

ISL.ts_invariant_statistical_loss — Method

ts_invariant_statistical_loss(rec, gen, Xₜ, Xₜ₊₁, hparams)

Train a model for time series data with statistical invariance loss method.

#Arguments

rec: The recurrent neural network (RNN) responsible for encoding the time series data.

gen: The generative model used for generating future time series data.

Xₜ: An array of input time series data at time t.

Xₜ₊₁: An array of target time series data at time t+1.

hparams::NamedTuple: A structure containing hyperparameters for training. It should include the following fields:

η::Float64: Learning rate for optimization.

window_size::Int: Size of the sliding window used during training.

K::Int: Number of samples in the generative model.

noise_model: Noise model used for generating random noise.

#Returns

losses::Vector{Float64}: A vector containing the training loss values for each iteration.

#Description This function train a model for time series data with statistical invariance loss method. It utilizes a recurrent neural network (rec) to encode the time series data at time t and a generative model (gen) to generate future time series data at time t+1. The training process involves optimizing both the rec and gen models.

The function iterates through the provided time series data (Xₜ and Xₜ₊₁) in batches, with a sliding window of size window_size.

ISL.γ — Method

γ(yₖ::Matrix{T}, yₙ::T, m::Int64) where {T<:AbstractFloat}

Calculate the contribution of ψₘ ∘ ϕ(yₖ, yₙ) to the m bin of the histogram as a Vector{Float}.

Arguments

yₖ::Matrix{T}: A matrix of values for which to compute the contribution.

yₙ::T: The center value around which the sigmoid and bump functions are centered.

m::Int64: The bin index for which to calculate the contribution.

Returns

A vector of floating-point values representing the contribution to the m bin of the histogram.

This function calculates the contribution of the composition of ψₘ and ϕ(yₖ, yₙ) to the m-th bin of the histogram. The result is a vector of floating-point values.

The contribution is computed according to the formula:

\[γ(yₖ, yₙ, m) = ψₘ ∘ ϕ(yₖ, yₙ)\]

ISL.γ_fast — Method

γ_fast(yₖ::Matrix{T}, yₙ::T, m::Int64) where {T<:AbstractFloat}`

Apply the γ function to the given parameters using StaticArrays for improved performance.

Arguments

yₖ::Matrix{T}: A matrix of values for which to compute the contribution.

yₙ::T: The center value around which the sigmoid and bump functions are centered.

m::Int64: The bin index for which to calculate the contribution.

Returns

A StaticVector{T} representing the contribution to the m bin of the histogram.

This function applies the γ function to compute the contribution of the composition of ψₘ and ϕ(yₖ, yₙ) to the m-th bin of the histogram. The result is a StaticVector{T} for improved performance.

Please note that although this function offers improved performance, it cannot be used in the training process with Zygote because Zygote does not support StaticArrays.

ISL.ψₘ — Method

ψₘ(y::T, m::Int64) where {T<:AbstractFloat}`

Calculate the bump function centered at m, implemented as a Gaussian function.

Arguments

y::T: The input value for which to compute the bump function.

m::Int64: The center point around which the bump function is centered.

Returns

A floating-point value representing the bump function's value at the input y.

This function calculates the bump function, which is centered at the integer value m. It is implemented as a Gaussian function with a standard deviation of 0.1.

ISL.ϕ — Method

ϕ(yₖ::Matrix{T}, yₙ::T) where {T<:AbstractFloat}

Calculate the sum of sigmoid functions centered at yₙ applied to the vector yₖ.

Arguments

yₖ::Matrix{T}: A matrix of values for which to compute the sum of sigmoid functions.

yₙ::T: The center value around which the sigmoid functions are centered.

Returns

A floating-point value representing the sum of sigmoid-transformed values.

This function calculates the sum of sigmoid functions, each centered at the value yₙ, applied element-wise to the matrix yₖ. The sum is computed according to the formula:

math ϕ(yₖ, yₙ) = ∑_{i=1}^K σ(yₖ^i, yₙ)`